+++ ts/connector-flowchart/flowchart-connector.ts
@@ -312,7 +312,7 @@ export class FlowchartConnector extends AbstractConnector {
                         comparator = pi["is" + axis.toUpperCase() + "GreaterThanStubTimes2"]
 
                     if (params.sourceEndpoint.elementId === params.targetEndpoint.elementId) {
-                        let _val = oss + ((1 - params.sourceEndpoint._anchor[otherAxis]) * params.sourceInfo[dim]) + this.maxStub
+                        let _val = oss + ((1 - params.sourceEndpoint._anchor.computedPosition[otherAxis]) * params.sourceInfo[dim]) + this.maxStub
                         return {
                             "x": [
                                 [ss, _val],